+++ scripts/Groovy/github/EnableBranchProtection.groovy
@@ -1,4 +1,3 @@
-@Grab(group = 'org.kohsuke', module = 'github-api', version = '1.93')
 @Grab(group = 'org.springframework', module = 'spring-webflux', version = '5.3.2')
 @Grab(group = 'io.projectreactor.netty', module = 'reactor-netty', version = '0.9.15.RELEASE')
 
@@ -6,7 +5,6 @@ import groovy.json.JsonOutput
 import groovy.transform.Field
 import groovy.cli.commons.CliBuilder
 
-import org.kohsuke.github.*
 import org.springframework.web.reactive.function.client.WebClient
 import org.springframework.web.reactive.function.client.ClientResponse
 
@@ -16,19 +14,24 @@ System.setProperty('java.util.logging.SimpleFormatter.format',
         '%1$tY-%1$tm-%1$tdT%1$tH:%1$tM:%1$tS.%1$tL%1$tz %4$s %5$s%6$s%n')
 @Field Logger LOGGER = Logger.getLogger('EnableBranchProtection.log')
 
-def cli = new CliBuilder(usage: 'groovy EnableBranchProtection.groovy -[options]')
+def cli = new CliBuilder(usage: 'groovy EnableBranchProtection.groovy -[options]', width: 200)
 cli.t(longOpt: 'token', args: 1, argName: 'token', defaultValue: 'GIT_TOKEN', 'Github access token')
-cli.n(longOpt: 'username', args: 1, argName: 'username', required: true, 'Github username corresponding to the access token')
+cli.n(longOpt: 'username', args: 1, argName: 'username', 'Github username corresponding to the access token')
 cli.a(longOpt: 'api', args: 1, argName: 'api', defaultValue: 'https://api.github.com', 'Git API URL')
 cli.o(longOpt: 'org', args: 1, argName: 'org', 'Github organization')
 cli.r(longOpt: 'repo', args: 1, argName: 'repo', required: true, 'Repository to which to enable branch protection')
 cli.b(longOpt: 'branches', args: 1, argName: 'branches', defaultValue: 'master,main',
         'Comma separated list of branches to protect')
 cli.s(longOpt: 'status-checks', args: 1, argName: 'status-checks', 'Comma separated list of required status checks')
+cli.rc(longOpt: 'review-count', args: 1, argName: 'review-count', defaultValue: '1', 'Number of reviews required')
+cli._(longOpt: 'help', args: 0, argName: 'help', 'Displays script usage instructions')
 
 def options = cli.parse(args)
 if (!options) {
     System.exit(1)
+} else if (options.help) {
+    cli.usage()
+    System.exit(0)
 }
 
 String gitUrl = options.a
@@ -36,12 +39,7 @@ def gitOrg = options.o
 String gitToken = System.getenv(options.t) ?: options.t
 String username = options.n
 String repoName = options.r
-
-GitHub github = GitHub.connectToEnterpriseWithOAuth(gitUrl, username, gitToken)
-
-// If org is supplied, look for the repo within the org. Else look for the repo within the supplied user name
-GHRepository repository = gitOrg ? github.getOrganization(gitOrg).getRepository(repoName) :
-        github.getUser(username).getRepository(repoName)
+int reviewCount = Integer.parseInt(options.rc)
 
 WebClient httpClient = WebClient.builder()
         .baseUrl(gitUrl)
@@ -53,9 +51,18 @@ def requiredStatusChecks = options.s ? [
         'contexts': options.s.split(',') as List
 ]: null
 
-repository.branches.entrySet().findAll { relevantBranches.contains(it.key) }.each { entry ->
+def branches = httpClient.get()
+        .uri("/repos/${gitOrg ?: username}/${repoName}/branches")
+        .header('Authorization', "token ${gitToken}")
+        .header('accept', 'application/vnd.github.v3+json')
+        .retrieve()
+        .bodyToMono(List)
+        .block()
+
+branches.findAll { relevantBranches.contains(it['name']) }.each { branch ->
     def requiredPullRequestReviews = [
-            "dismiss_stale_reviews": true
+            "dismiss_stale_reviews": true,
+            "required_approving_review_count": reviewCount
     ]
 
     def dismissalRestrictions = null
@@ -76,7 +83,7 @@ repository.branches.entrySet().findAll { relevantBranches.contains(it.key) }.eac
     LOGGER.info({ JsonOutput.toJson(requestBody) })
 
     ClientResponse response = httpClient.put()
-            .uri("/repos/${gitOrg ?: username}/${repoName}/branches/${entry.key}/protection")
+            .uri("/repos/${gitOrg ?: username}/${repoName}/branches/${branch['name']}/protection")
             .header('Authorization', "token ${gitToken}")
             .header('Content-Type', 'application/json')
             .header('accept', 'application/vnd.github.v3+json')